# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_transport_http_announcement.AdminAnnouncementListResponseDoc:
    properties:
      data:
        properties:
          results:
            items:
              $ref: '#/definitions/internal_transport_http_announcement.AnnouncementResponse'
            type: array
          total:
            type: integer
        type: object
      errorMsg:
        type: string
    type: object
  internal_transport_http_announcement.AnnouncementCloseDataResponse:
    properties:
      closed:
        type: boolean
    type: object
  internal_transport_http_announcement.AnnouncementCloseResponseDoc:
    properties:
      data:
        $ref: '#/definitions/internal_transport_http_announcement.AnnouncementCloseDataResponse'
      errorMsg:
        type: string
    type: object
  internal_transport_http_announcement.AnnouncementDataResponse:
    properties:
      announcement:
        $ref: '#/definitions/internal_transport_http_announcement.AnnouncementResponse'
    type: object
  internal_transport_http_announcement.AnnouncementDeleteDataResponse:
    properties:
      deleted:
        type: boolean
    type: object
  internal_transport_http_announcement.AnnouncementDeleteResponseDoc:
    properties:
      data:
        $ref: '#/definitions/internal_transport_http_announcement.AnnouncementDeleteDataResponse'
      errorMsg:
        type: string
    type: object
  internal_transport_http_announcement.AnnouncementDismissDataResponse:
    properties:
      dismissed:
        type: boolean
    type: object
  internal_transport_http_announcement.AnnouncementDismissResponseDoc:
    properties:
      data:
        $ref: '#/definitions/internal_transport_http_announcement.AnnouncementDismissDataResponse'
      errorMsg:
        type: string
    type: object
  internal_transport_http_announcement.AnnouncementListResponseDoc:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_transport_http_announcement.AnnouncementResponse'
        type: array
      errorMsg:
        type: string
    type: object
  internal_transport_http_announcement.AnnouncementResponse:
    properties:
      closedAt:
        type: string
      contentMarkdown:
        type: string
      createdAt:
        type: string
      createdByUserID:
        type: integer
      expiresAt:
        type: string
      id:
        type: integer
      pinned:
        type: boolean
      priority:
        type: integer
      startsAt:
        type: string
      status:
        type: string
      title:
        type: string
      type:
        type: string
      updatedAt:
        type: string
    type: object
  internal_transport_http_announcement.AnnouncementResponseDoc:
    properties:
      data:
        $ref: '#/definitions/internal_transport_http_announcement.AnnouncementDataResponse'
      errorMsg:
        type: string
    type: object
  internal_transport_http_announcement.AnnouncementStateRequest:
    properties:
      updatedAt:
        type: string
    required:
    - updatedAt
    type: object
  internal_transport_http_announcement.CreateAnnouncementRequest:
    properties:
      contentMarkdown:
        maxLength: 20000
        minLength: 1
        type: string
      expiresAt:
        type: string
      pinned:
        type: boolean
      priority:
        type: integer
      startsAt:
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
      title:
        maxLength: 120
        minLength: 1
        type: string
      type:
        enum:
        - critical
        - warning
        - info
        - normal
        - general
        type: string
    required:
    - contentMarkdown
    - title
    type: object
  internal_transport_http_announcement.ErrorDoc:
    properties:
      data: {}
      details: {}
      errorCode:
        example: invalid_request
        type: string
      errorMsg:
        example: invalid request
        type: string
      requestId:
        example: ""
        type: string
    type: object
  internal_transport_http_announcement.PatchAnnouncementRequestDoc:
    properties:
      contentMarkdown:
        maxLength: 20000
        type: string
      expiresAt:
        type: string
      pinned:
        type: boolean
      priority:
        type: integer
      startsAt:
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
      title:
        maxLength: 120
        type: string
      type:
        enum:
        - critical
        - warning
        - info
        - normal
        - general
        type: string
    type: object
...
info:
  contact: {}
  description: DEEIX Chat 后端 API 文档
  title: DEEIX Chat API
  version: "0.2.4"
paths:
  /admin/announcements:
    get:
      consumes:
      - application/json
      description: 分页查询站点公告
      parameters:
      - description: 状态：active/inactive
        in: query
        name: status
        type: string
      - description: 类型：critical/warning/info/normal/general
        in: query
        name: type
        type: string
      - description: 是否置顶
        in: query
        name: pinned
        type: boolean
      - description: 搜索关键词
        in: query
        name: q
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_announcement.AdminAnnouncementListResponseDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_announcement.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 管理员查询公告
      tags:
      - admin-announcements
    post:
      consumes:
      - application/json
      description: 创建一条 Markdown 站点公告
      parameters:
      - description: 公告配置
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_announcement.CreateAnnouncementRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_announcement.AnnouncementResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_announcement.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_announcement.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 管理员创建公告
      tags:
      - admin-announcements
  /admin/announcements/{id}:
    delete:
      consumes:
      - application/json
      description: 软删除公告
      parameters:
      - description: 公告ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_announcement.AnnouncementDeleteResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_announcement.ErrorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http_announcement.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_announcement.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 管理员删除公告
      tags:
      - admin-announcements
    patch:
      consumes:
      - application/json
      description: 更新公告标题、内容、状态、优先级和有效期
      parameters:
      - description: 公告ID
        in: path
        name: id
        required: true
        type: integer
      - description: 公告更新字段
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_announcement.PatchAnnouncementRequestDoc'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_announcement.AnnouncementResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_announcement.ErrorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http_announcement.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_announcement.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 管理员更新公告
      tags:
      - admin-announcements
...
  /announcements:
    get:
      consumes:
      - application/json
      description: 登录用户获取当前可展示的站点公告列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_announcement.AnnouncementListResponseDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_announcement.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 获取当前公告
      tags:
      - announcements
  /announcements/{id}/close:
    post:
      consumes:
      - application/json
      description: 登录用户关闭当前公告版本；公告更新后会重新展示
      parameters:
      - description: 公告ID
        in: path
        name: id
        required: true
        type: integer
      - description: 公告版本
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_announcement.AnnouncementStateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_announcement.AnnouncementCloseResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_announcement.ErrorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http_announcement.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_announcement.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 关闭公告
      tags:
      - announcements
  /announcements/{id}/dismiss-today:
    post:
      consumes:
      - application/json
      description: 登录用户对当前公告版本记录今日不再显示
      parameters:
      - description: 公告ID
        in: path
        name: id
        required: true
        type: integer
      - description: 公告版本
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_announcement.AnnouncementStateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_announcement.AnnouncementDismissResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_announcement.ErrorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http_announcement.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_announcement.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 今日不再显示公告
      tags:
      - announcements
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"